+++ src/components/Timeline/PostsBox.js
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import TimelineMessage from "../../assets/styles/TimelineMessage";
 import PostCard from "./PostCard";
 
-export default function PostsBox({ setRefresh, posts, userEmail }) {
+export default function PostsBox({ refresh, setRefresh, posts, userEmail }) {
   return (
     <Wrapper>
       {posts.length > 0 ? (
@@ -16,6 +16,7 @@ export default function PostsBox({ setRefresh, posts, userEmail }) {
               username={post.name}
               postDescription={post.description}
               link={post.link}
+              refresh={refresh}
               setRefresh={setRefresh}
               userEmail={userEmail}
             />
