+++ src/components/Timeline/Timeline.js
@@ -39,11 +39,16 @@ export default function Timeline() {
       <Title>timeline</Title>
       <PublishBox>
         <ProfilePic src={userImage} />
-        <FormBox updatePosts={setPosts} />
+        <FormBox
+          refresh={refresh}
+          setRefresh={setRefresh}
+          updatePosts={setPosts}
+        />
       </PublishBox>
       <Posts>
         {posts ? (
           <PostsBox
+            refresh={refresh}
             setRefresh={setRefresh}
             userEmail={userEmail}
             posts={posts}
@@ -71,36 +76,36 @@ const Wrapper = styled.div`
   }
 
   @media (max-width: 614px) {
-        width: 100vw;
-        margin-top: 30px;
-        h2{
-          font-size: 40px;
-          margin-left: 10px;
-        }
+    width: 100vw;
+    margin-top: 30px;
+    h2 {
+      font-size: 40px;
+      margin-left: 10px;
     }
+  }
 `;
 
-const PublishBox = styled.div`  
-    margin-top: 60px;
-    margin-bottom: 15px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: var(--main-white);
-    height: fit-content;
-    width: 40vw;
-    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-    border-radius: 16px;
-    img {
-        display: none;
-    }
+const PublishBox = styled.div`
+  margin-top: 60px;
+  margin-bottom: 15px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: var(--main-white);
+  height: fit-content;
+  width: 40vw;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  border-radius: 16px;
+  img {
+    display: none;
+  }
 
-    @media (max-width: 614px) {
-      width: 100vw;
-      margin-top: 35px;
-      border-radius: 0px;
-    }
-`
+  @media (max-width: 614px) {
+    width: 100vw;
+    margin-top: 35px;
+    border-radius: 0px;
+  }
+`;
 
 const Posts = styled.div`
   margin-top: 15px;
@@ -111,6 +116,6 @@ const Posts = styled.div`
   align-items: center;
 
   @media (max-width: 614px) {
-        width: 100vw;
-    }
+    width: 100vw;
+  }
 `;
